+++ cli/spiff/f2_handlers.cpp
@@ -1,8 +1,8 @@
 #include "f2_handlers.h"
 #include "bit_t.h"
-#include "cmdlinematops.h"
-#include "cmdlinevecops.h"
-#include "f2linalg.h"
+#include "cmd_line_matrix_ops.h"
+#include "cmd_line_vector_ops.h"
+#include "f2_linear_algebra.h"
 #include "f2poly_t.h"
 #include "int_random.h"
 #include "tmatrix.h"
