+++ cli/spiff/f2npoly_handlers.cpp
@@ -1,12 +1,12 @@
 #include "f2npoly_handlers.h"
-#include "cmdlinematops.h"
-#include "cmdlineops.h"
-#include "cmdlinevecops.h"
+#include "cmd_line_matrix_ops.h"
+#include "cmd_line_ops.h"
+#include "cmd_line_vector_ops.h"
 #include "f2npoly_factor.h"
 #include "f2npoly_random.h"
 #include "f2npoly_t.h"
 #include "f2npoly_totient.h"
-#include "f2pmlinalg.h"
+#include "f2_polymod_linear_algebra.h"
 #include "qff.h"
 #include "qpoly_t.h"
 #include "sp_list_elts.h"
