+++ cli/spiff/f2polyrat_handlers.cpp
@@ -1,7 +1,7 @@
 #include "f2polyrat_handlers.h"
-#include "cmdlinematops.h"
-#include "cmdlineops.h"
-#include "cmdlinevecops.h"
+#include "cmd_line_matrix_ops.h"
+#include "cmd_line_ops.h"
+#include "cmd_line_vector_ops.h"
 #include "f2poly_random.h"
 #include "f2polyrat_t.h"
