+++ cli/spiff/fp_handlers.cpp
@@ -1,9 +1,9 @@
 #include "fp_handlers.h"
-#include "cmdlinematops.h"
-#include "cmdlineops.h"
-#include "cmdlinevecops.h"
+#include "cmd_line_matrix_ops.h"
+#include "cmd_line_ops.h"
+#include "cmd_line_vector_ops.h"
 #include "fp_units.h"
-#include "fplinalg.h"
+#include "fp_linear_algebra.h"
 #include "intmod_random.h"
 #include "intmod_t.h"
 #include "sp_list_elts.h"
