+++ cli/spiff/fppoly_handlers.cpp
@@ -1,8 +1,8 @@
 #include "fppoly_handlers.h"
-#include "cmdlinematops.h"
-#include "cmdlineops.h"
-#include "cmdlinevecops.h"
-#include "fplinalg.h"
+#include "cmd_line_matrix_ops.h"
+#include "cmd_line_ops.h"
+#include "cmd_line_vector_ops.h"
+#include "fp_linear_algebra.h"
 #include "fppoly_factor.h"
 #include "fppoly_random.h"
 #include "fppoly_t.h"
