+++ cli/spiff/fppolymod_handlers.cpp
@@ -1,13 +1,13 @@
 #include <string.h>
 
-#include "cmdlinematops.h"
-#include "cmdlineops.h"
-#include "cmdlinevecops.h"
+#include "cmd_line_matrix_ops.h"
+#include "cmd_line_ops.h"
+#include "cmd_line_vector_ops.h"
 #include "fppolymod_handlers.h"
 #include "fppolymod_random.h"
 #include "fppolymod_t.h"
 #include "sp_list_elts.h"
-#include "spminchar.h"
+#include "minimal_and_characteristic_polynomials.h"
 #include "tbl_types.h"
 
 void fppmlist_usage(char *argv0) {
