+++ pre-cmake/tools/f2pmop.cpp
@@ -8,7 +8,7 @@
 #include <stdio.h>
 
 #include "f2polymod_t.h"
-#include "cmdlineops.h"
+#include "cmd_line_ops.h"
 
 typedef void usage_t(char * argv0);
 typedef int main_t (int argc, char ** argv, usage_t * pusage);
