+++ pre-cmake/tools/zop.cpp
@@ -7,7 +7,7 @@
 #include <iostream>
 #include <stdio.h>
 
-#include "cmdlineops.h"
+#include "cmd_line_ops.h"
 
 typedef void usage_t(char * argv0);
 typedef int main_t (int argc, char ** argv, usage_t * pusage);
