+++ src/linalg/f2_linear_algebra.cpp
@@ -4,7 +4,7 @@
 // Please see LICENSE.txt.
 // ================================================================
 
-#include "f2linalg.h"
+#include "f2_linear_algebra.h"
 #include "f2poly_t.h"
 
 // ----------------------------------------------------------------
