+++ src/linalg/minimal_and_characteristic_polynomials.h
