+++ src/units/f2polymod_convert.cpp
@@ -5,7 +5,7 @@
 // ================================================================
 
 #include "f2polymod_convert.h"
-#include "f2linalg.h"
+#include "f2_linear_algebra.h"
 #include "f2npoly_factor.h"
 #include "f2polymod_units.h"
